app/requirements.txt
black==23.11.0
Django==4.2.7
django-admin-interface==0.28.6
django-ninja==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyTelegramBotAPI==4.16.1
pytz==2023.3.post1
sentry-sdk==1.39.1
requests==2.31.0
Pillow==10.1.0
django-model-utils==4.3.1
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1